DiagnosticSettings as switch parameter
From a powershell perspective, it makes most sense that when we have bool parameters that are default false, they should be of the type SwitchParameter.

This enables the user to just add the parameter as a switch, instead of having to add the parameter and parse $true to them.

E.g.
-DiagnosticSettings:$true
becomes
-DiagnosticSettings

This implementation should be more consistent with the powershell paradigm